requirements.txt
PySide6
pyqtdarktheme
Pillow
scikit-image
torch
lpips
numpy
